nextg3n/requirements.txt
# Core dependencies
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Machine learning dependencies
torch>=2.0.0
pytorch-lightning>=2.0.0
pytorch-forecasting>=1.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
faiss-gpu>=1.7.2
modAL>=0.4.1

# Trading dependencies
alpaca-trade-api>=3.0.0
backtrader>=1.9.76.123

# Messaging dependencies
kafka-python>=2.0.2

# Storage dependencies
redis>=4.5.0
chromadb>=0.4.0

# Web framework dependencies
flask==2.2.5

# Visualization dependencies
matplotlib>=3.7.2
plotly>=5.14.0
pandas>=2.0.0
numpy>=1.24.0

# Networking dependencies
aiohttp>=3.8.0
requests>=2.28.2

# Testing dependencies
pytest>=7.3.0
unittest-mock>=1.0.0

# Automation dependencies
pyautogen>=0.2.0

# Scheduling dependencies
apscheduler>=3.10.1
pandas_market_calendars>=4.4.0

# Alerting dependencies
slack_sdk>=3.27.0